requirements-dev.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.5
django-filter==1.0.4
djangorestframework==3.6.4
google-api-python-client==1.6.2
httplib2==0.10.3
Markdown==2.6.9
oauth2client==4.1.2
pyasn1==0.3.1
pyasn1-modules==0.0.10
python-decouple==3.1
pytz==2017.2
rsa==3.4.2
six==1.10.0
static3==0.7.0
uritemplate==3.0.0
vboxapi==1.0